Continuacao
Tutorial 18,19, 20. Tudo funcional
Tentar introduzir fadeIn na cena Informacao.
Vou continuar foi so para guardar para nao perder nada

+++ Assets/Scripts/CollectableRotate.cs
@@ -6,7 +6,7 @@ using System.Collections;
 public class CollectableRotate : MonoBehaviour
 {
     
-    [SerializeField] int rotationSpeed = 1;
+    [SerializeField] float rotationSpeed = 1;
 
     void Update()
     {
